feat: NavBottom 이벤트 기능 추가 완료
