+++ src/lib/components/NavigationBar/Bottom/BottomItem/styles.ts
