+++ src/lib/components/NavigationBar/type.ts
@@ -1,3 +1,17 @@
 export interface INavBar {
-  page?: 'home' | 'reserve' | 'member' | 'center' | 'mypage';
+  children?: React.ReactNode;
+  memberName?: string;
+}
+
+export interface INavBottom {
+  onClickNavItem: (arg0: string) => void;
+  selected: string;
+  isLabel: boolean;
+}
+
+export interface INavBottomItem extends React.HTMLAttributes<HTMLElement> {
+  icon: string;
+  label: string;
+  isSelected: boolean;
+  isLabel: boolean;
 }
